• move defaults registration from window into seperate command
This is both because future commands (like file requesters with auxilary views) can benefit from this and to offer more (sub)commands related to the defaults.

Presently there is a --read «key» in addition to --register «dictionary». If the key is not found, no output is written back, ideally "$DIALOG" should exit with a non-zero return code, but our infrastructure does not currently allow that.


git-svn-id: http://svn.textmate.org/branches/WIP/Tools/Dialog2@10880 dfb7d73b-c2ec-0310-8fea-fb051d288c6d