Adjust pop-up menu width to widest row
Previously we would go by string length and multiply it by 18 (average width of a character?) to get the width. The width would not be allowed to exceed 340, this has now been increased to 600. Furthermore, the width would never decrease, unsure if this was to limit visual distractions or if it was because the width was a heuristic, so there was a chance that the calculated width was too small to have the rows fit, and by newer shrinking the menu’s width, we would decrease the chance of this happening.