Replace deprecated method for initializing nibs (10.8)
The new method requires that the nib be compiled. If the nib is a file package, we try to load the enclosed `keyedobjects.nib` file, otherwise we just use the path as is. Note that we still assume that file contains the .nib extension so when compiling a .xib file the extension should still be .nib (which is the usual convention). To compile .xib use

```
ibtool --compile myNib.nib myNib.xml
```