Determine pop-up menu width using cellSize instead of cellSizeForBounds:
The latter method does not take the potential image width into account. The proper fix would be to extend the TMDImageAndTextCell to account for image in an overloaded cellSizeForBounds: method but that is more work and cells are deprecated, so I see little reason to improve code that should be replaced.